requirements.txt
torch==2.1.0
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.2
xgboost==2.0.0
lightgbm==4.1.0
imbalanced-learn==0.11.0
openml==0.14.1
pyyaml==6.0.1
scipy==1.11.3
matplotlib==3.8.0
seaborn==0.13.0
tqdm==4.66.1
pydantic==2.4.2
pytest==7.4.2
